chore(quality): add ruff gate, extract config_loading + stats modules
Addresses review recs #2 and the safe half of #5.

#2 — Ruff is the static-check gate impl_loop.py ships as the default for
consumers, but it had never been run on this repo (44 issues). Add a ruff
config to pyproject.toml, clear all 44 findings (unused imports, dead
locals, an ambiguous name, a semicolon statement), and add a CI workflow
that runs ruff + the full pytest suite on every PR — the repo's first
lint/test gate.

#5 (safe extractions) — pull the two lowest-risk seams out of the
3,066-line run_phase.py:
- config_loading.py: the tomllib/tomli fallback was copy-pasted across 6
  files with divergent behavior (some exited friendly, some crashed raw).
  One shared copy, friendly everywhere.
- stats.py: the pure cross-run aggregation/formatting (aggregate_stats,
  format_stats, _summarize_metrics, _parse_usage_log). The I/O handlers
  stay in run_phase.py, so no import cycle.

run_phase.py: 3,066 -> 2,834 lines. 624 tests green, ruff clean.
The remaining run_phase.py split is deferred with a detailed plan (Fable 5).
