+++ .github/workflows/ci.yml
@@ -1,0 +1,38 @@
+name: CI
+
+on:
+  pull_request:
+    branches: [main]
+  push:
+    branches: [main]
+  merge_group:
+
+jobs:
+  lint-and-test:
+    name: Lint and test
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: studio
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install ruff pytest
+          # Runtime dep for the tomli fallback path (harmless on 3.11+).
+          pip install "tomli>=2.0; python_version < '3.11'"
+
+      # ruff is the same static-check gate impl_loop.py ships as the default
+      # for consumers, so we run it on ourselves here.
+      - name: Ruff
+        run: ruff check .
+
+      - name: Tests
+        run: python -m pytest tests/ -q
