+++ studio/integrations/slack_digest.py
@@ -29,10 +29,7 @@ problem can never break the run it is reporting on.
 """
 from __future__ import annotations
 
-try:
-    import tomllib  # Python 3.11+
-except ModuleNotFoundError:
-    import tomli as tomllib  # type: ignore[no-redefine]  # Python 3.10 fallback
+from config_loading import tomllib
 
 import json
 import os
