+++ studio/pyproject.toml
@@ -11,3 +11,15 @@ dependencies = [
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[tool.ruff]
+target-version = "py310"
+line-length = 100
+# The generated-artifact and cache dirs are not source.
+extend-exclude = ["output", "knowledge", ".studio", ".scratch"]
+
+[tool.ruff.lint]
+# Default rules (pyflakes + a slice of pycodestyle). This is the same gate
+# `impl_loop.py` ships as the default static check for consumers, so we hold
+# ourselves to it too.
+select = ["E4", "E7", "E9", "F"]
