+++ studio/role_overrides.py
@@ -13,7 +13,7 @@ from __future__ import annotations
 
 import json
 from pathlib import Path
-from typing import Dict, List, Optional
+from typing import Dict, Optional
 
 
 ROLES_DIRNAME = "roles"
