+++ studio/tests/test_run_phase_roles.py
@@ -1,8 +1,6 @@
 """Unit tests for run_phase_roles.py."""
 from __future__ import annotations
 
-import json
-from pathlib import Path
 
 import pytest
